Add files via upload
Notes: The Steem users use the term "Approved at 75%" for example followed by their comment (should be critical) and should have their Steem account linked to their(s) version control system account.
Second notes: Further mathematical and science-based description of the protocol are to come and can largely vary.
Disclaimer: This is a right of the public domain, not any members of the team should be taken responsible for any use of such advanced protocol. This is given for free and for everyone, we own the right of the publication.